requirements.txt
Django==3.2.9
Pillow==8.3.1
djangorestframework==3.13.0
drf-spectacular==0.21.0
django-filter== 21.1